+++ pyccel/parser/semantic.py
@@ -270,8 +270,21 @@ class SemanticParser(BasicParser):
     #                     Public functions
     #================================================================
 
-    def annotate(self, **settings):
-        """."""
+    def annotate(self):
+        """
+        Add type information to the AST.
+
+        This function is the entry point for this class. It annotates the
+        AST object created by the syntactic stage which was collected
+        in the constructor. The annotation adds all necessary information
+        about the type etc to describe the object sufficiently well for
+        printing. See the developer docs for more details.
+
+        Returns
+        -------
+        pyccel.ast.basic.Basic
+            An annotated object which can be printed.
+        """
 
         if self.semantic_done:
             print ('> semantic analysis already done')
@@ -292,7 +305,7 @@ class SemanticParser(BasicParser):
         self._allocs.append([])
         # we add the try/except to allow the parser to find all possible errors
         pyccel_stage.set_stage('semantic')
-        ast = self._visit(ast, **settings)
+        ast = self._visit(ast)
 
         self._ast = ast
 
@@ -464,9 +477,31 @@ class SemanticParser(BasicParser):
         self._allocs.pop()
         return deallocs
 
-    def _infer_type(self, expr, **settings):
+    def _infer_type(self, expr):
         """
-        type inference for expressions
+        Infer all relevant type information for the expression.
+
+        Create a dictionary describing all the type information that can be
+        inferred about the expression `expr`. This includes information about:
+        - `datatype`
+        - `precision`
+        - `rank`
+        - `shape`
+        - `order`
+        - `memory_handling`
+        - `cls_base`
+        - `is_target`
+
+        Parameters
+        ----------
+        expr : pyccel.ast.basic.Basic
+                An AST object representing an object in the code whose type
+                must be determined.
+
+        Returns
+        -------
+        dict
+            Dictionary containing all the type information which was inferred.
         """
         # TODO - add settings to Errors
         #      - line and column
@@ -474,16 +509,12 @@ class SemanticParser(BasicParser):
 
         errors = Errors()
 
-        verbose = settings.pop('verbose', False)
-        if verbose:
-            print ('*** type inference for : ', type(expr))
-
         d_var = {}
         # TODO improve => put settings as attribut of Parser
 
         if expr in (PythonInt, PythonFloat, PythonComplex, PythonBool, NumpyBool, NumpyInt, NumpyInt8, NumpyInt16,
                       NumpyInt32, NumpyInt64, NumpyComplex, NumpyComplex64,
-					  NumpyComplex128, NumpyFloat, NumpyFloat64, NumpyFloat32):
+                      NumpyComplex128, NumpyFloat, NumpyFloat64, NumpyFloat32):
 
             d_var['datatype'   ] = '*'
             d_var['rank'       ] = 0
@@ -525,7 +556,7 @@ class SemanticParser(BasicParser):
             return d_var
 
         elif isinstance(expr, Duplicate):
-            d = self._infer_type(expr.val, **settings)
+            d = self._infer_type(expr.val)
 
             # TODO must check that it is consistent with pyccel's rules
             # TODO improve
@@ -753,7 +784,7 @@ class SemanticParser(BasicParser):
             else:
                 return PythonTuple(*(val.args*length))
 
-    def _handle_function_args(self, arguments, **settings):
+    def _handle_function_args(self, arguments):
         """
         Get a list of all function arguments.
 
@@ -768,9 +799,6 @@ class SemanticParser(BasicParser):
         arguments : list of FunctionCallArgument
             The arguments which were passed to the function.
 
-        **settings : dict
-            Any settings to be passed to the `_visit` function.
-
         Returns
         -------
         list of FunctionCallArgument
@@ -778,7 +806,7 @@ class SemanticParser(BasicParser):
         """
         args  = []
         for arg in arguments:
-            a = self._visit(arg, **settings)
+            a = self._visit(arg)
             if isinstance(a.value, StarredArguments):
                 args.extend([FunctionCallArgument(av) for av in a.value.args_var])
             else:
@@ -853,7 +881,7 @@ class SemanticParser(BasicParser):
                         symbol = expr,
                         severity='error')
 
-    def _handle_function(self, expr, func, args, **settings):
+    def _handle_function(self, expr, func, args):
         """
         Create the node representing the function call.
 
@@ -870,9 +898,6 @@ class SemanticParser(BasicParser):
 
         args : tuple
                The arguments passed to the function.
-
-        **settings : dict
-            The settings passed to _visit functions.
 
         Returns
         -------
@@ -1037,40 +1062,44 @@ class SemanticParser(BasicParser):
             d_lhs['memory_handling'] = 'alias'
             rhs.base.is_target = not rhs.base.is_alias
 
-    def _assign_lhs_variable(self, lhs, d_var, rhs, new_expressions, is_augassign,arr_in_multirets=False, **settings):
+    def _assign_lhs_variable(self, lhs, d_var, rhs, new_expressions, is_augassign,arr_in_multirets=False):
         """
-        Create a lhs based on the information in d_var
-        If the lhs already exists then check that it has the expected properties.
+        Create a variable from the left-hand side (lhs) of an assignment.
+        
+        Create a lhs based on the information in d_var, if the lhs already exists
+        then check that it has the expected properties.
 
         Parameters
         ----------
         lhs : PyccelSymbol (or DottedName of PyccelSymbols)
-            The representation of the lhs provided by the SyntacticParser
+            The representation of the lhs provided by the SyntacticParser.
 
         d_var : dict
-            Dictionary of expected lhs properties
+            Dictionary of expected lhs properties.
 
         rhs : Variable / expression
             The representation of the rhs provided by the SemanticParser.
             This is necessary in order to set the rhs 'is_target' property
-            if necessary
+            if necessary.
 
-        new_expression : list
+        new_expressions : list
             A list which allows collection of any additional expressions
-            resulting from this operation (e.g. Allocation)
+            resulting from this operation (e.g. Allocation).
 
         is_augassign : bool
             Indicates whether this is an assign ( = ) or an augassign ( += / -= / etc )
             This is necessary as the restrictions on the dtype are less strict in this
-            case
+            case.
 
         arr_in_multirets : bool
             If True, rhs has an array in its results, otherwise, it should be set to False.
             It helps when we don't need lhs to be a pointer in case of a returned array in
             a tuple of results.
 
-        settings : dictionary
-            Provided to all _visit_ClassName functions
+        Returns
+        -------
+        pyccel.ast.variable.Variable
+            The representation of the lhs provided by the SemanticParser.
         """
 
         if isinstance(lhs, IndexedElement):
@@ -1230,7 +1259,7 @@ class SemanticParser(BasicParser):
                 new_cls = ClassDef(cls_name, attributes, [], superclass=parent)
                 self.scope.parent_scope.insert_class(new_cls)
             else:
-                lhs = self._visit(lhs, **settings)
+                lhs = self._visit(lhs)
         else:
             lhs_type = str(type(lhs))
             raise NotImplementedError(f"_assign_lhs_variable does not handle {lhs_type}")
@@ -1407,21 +1436,24 @@ class SemanticParser(BasicParser):
         if var.precision == -1 and precision != var.precision:
             var.use_exact_precision()
 
-    def _assign_GeneratorComprehension(self, lhs_name, expr, **settings):
+    def _assign_GeneratorComprehension(self, lhs_name, expr):
         """
-        Visit the GeneratorComprehension node creating all necessary expressions
-        for its definition
+        Visit the GeneratorComprehension node.
+
+        Create all necessary expressions for the
+        GeneratorComprehension node definition.
 
         Parameters
         ----------
         lhs_name : str
-                    The name to which the expression is assigned
+                    The name to which the expression is assigned.
         expr : GeneratorComprehension
+                The GeneratorComprehension node.
 
-        Results
+        Returns
         -------
-        new_expr : CodeBlock
-                   CodeBlock containing the semantic version of the GeneratorComprehension node
+        pyccel.ast.functionalexpr.GeneratorComprehension
+                CodeBlock containing the semantic version of the GeneratorComprehension node
         """
         result   = expr.expr
 
@@ -1433,7 +1465,7 @@ class SemanticParser(BasicParser):
         new_expr = []
         while isinstance(loop, For):
             nlevels+=1
-            iterable = Iterable(self._visit(loop.iterable, **settings))
+            iterable = Iterable(self._visit(loop.iterable))
             n_index = max(1, iterable.num_loop_counters_required)
             # Set dummy indices to iterable object in order to be able to
             # obtain a target with a deducible dtype
@@ -1468,12 +1500,12 @@ class SemanticParser(BasicParser):
 
         # Visit result expression (correctly defined as iterator
         # objects exist in the scope despite not being defined)
-        result = self._visit(result, **settings)
+        result = self._visit(result)
         if isinstance(result, CodeBlock):
             result = result.body[-1]
 
         # Infer the final dtype of the expression
-        d_var = self._infer_type(result, **settings)
+        d_var = self._infer_type(result)
         dtype = d_var.pop('datatype')
         d_var['is_temp'] = expr.lhs.is_temp
 
@@ -1488,7 +1520,7 @@ class SemanticParser(BasicParser):
         # Iterate over the loops
         # This provides the definitions of iterators as well
         # as the central expression
-        loops = [self._visit(expr.loops, **settings)]
+        loops = [self._visit(expr.loops)]
 
         # If necessary add additional expressions corresponding
         # to nested GeneratorComprehensions
@@ -1530,14 +1562,25 @@ class SemanticParser(BasicParser):
     #====================================================
 
 
-    def _visit(self, expr, **settings):
-        """Annotates the AST.
+    def _visit(self, expr):
+        """
+        Annotate the AST.
 
         The annotation is done by finding the appropriate function _visit_X
         for the object expr. X is the type of the object expr. If this function
         does not exist then the method resolution order is used to search for
         other compatible _visit_X functions. If none are found then an error is
-        raised
+        raised.
+        
+        Parameters
+        ----------
+        expr : pyccel.ast.basic.Basic
+            Object to visit of type X.
+        
+        Returns
+        -------
+        pyccel.ast.basic.Basic
+            AST object which is the semantic equivalent of expr.
         """
 
         # TODO - add settings to Errors
@@ -1552,7 +1595,7 @@ class SemanticParser(BasicParser):
         for cls in classes:
             annotation_method = '_visit_' + cls.__name__
             if hasattr(self, annotation_method):
-                obj = getattr(self, annotation_method)(expr, **settings)
+                obj = getattr(self, annotation_method)(expr)
                 if isinstance(obj, Basic) and self._current_fst_node:
                     obj.set_fst(self._current_fst_node)
                 self._current_fst_node = current_fst
@@ -1563,7 +1606,7 @@ class SemanticParser(BasicParser):
             bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
             severity='fatal')
 
-    def _visit_Module(self, expr, **settings):
+    def _visit_Module(self, expr):
         body = self._visit(expr.program).body
         program_body      = []
         init_func_body    = []
@@ -1739,15 +1782,12 @@ class SemanticParser(BasicParser):
 
         return mod
 
-    def _visit_tuple(self, expr, **settings):
-        return tuple(self._visit(i, **settings) for i in expr)
-
-    def _visit_PythonTuple(self, expr, **settings):
-        ls = [self._visit(i, **settings) for i in expr]
+    def _visit_PythonTuple(self, expr):
+        ls = [self._visit(i) for i in expr]
         return PythonTuple(*ls)
 
-    def _visit_PythonList(self, expr, **settings):
-        ls = [self._visit(i, **settings) for i in expr]
+    def _visit_PythonList(self, expr):
+        ls = [self._visit(i) for i in expr]
         expr = PythonList(*ls)
 
         if not expr.is_homogeneous:
@@ -1756,8 +1796,8 @@ class SemanticParser(BasicParser):
                 severity='fatal')
         return expr
 
-    def _visit_FunctionCallArgument(self, expr, **settings):
-        value = self._visit(expr.value, **settings)
+    def _visit_FunctionCallArgument(self, expr):
+        value = self._visit(expr.value)
         a = FunctionCallArgument(value, expr.keyword)
         if isinstance(a.value, PyccelArithmeticOperator) and a.value.rank:
             tmp_var = self.scope.get_new_name()
@@ -1766,20 +1806,13 @@ class SemanticParser(BasicParser):
             a = FunctionCallArgument(self._visit(tmp_var))
         return a
 
-    def _visit_FunctionDefArgument(self, expr, **settings):
-        var   = self._visit(expr.var, **settings)
-        value = self._visit(expr.value, **settings)
-        return FunctionDefArgument(var, value=value,
-                annotation=expr.annotation,
-                kwonly=expr.is_kwonly)
-
-    def _visit_CodeBlock(self, expr, **settings):
+    def _visit_CodeBlock(self, expr):
         ls = []
         self._additional_exprs.append([])
         for b in expr.body:
 
             # Save parsed code
-            line = self._visit(b, **settings)
+            line = self._visit(b)
             ls.extend(self._additional_exprs[-1])
             self._additional_exprs[-1] = []
             if isinstance(line, CodeBlock):
@@ -1790,35 +1823,35 @@ class SemanticParser(BasicParser):
 
         return CodeBlock(ls)
 
-    def _visit_Nil(self, expr, **settings):
+    def _visit_Nil(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_EmptyNode(self, expr, **settings):
+    def _visit_EmptyNode(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_Break(self, expr, **settings):
+    def _visit_Break(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_Continue(self, expr, **settings):
+    def _visit_Continue(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_Comment(self, expr, **settings):
+    def _visit_Comment(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_CommentBlock(self, expr, **settings):
+    def _visit_CommentBlock(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_AnnotatedComment(self, expr, **settings):
+    def _visit_AnnotatedComment(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_OmpAnnotatedComment(self, expr, **settings):
+    def _visit_OmpAnnotatedComment(self, expr):
         code = expr._user_nodes
         code = code[-1]
         index = code.body.index(expr)
@@ -1853,7 +1886,7 @@ class SemanticParser(BasicParser):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_Omp_End_Clause(self, expr, **settings):
+    def _visit_Omp_End_Clause(self, expr):
         end_loop = any(c in expr.txt for c in ['for', 'distribute', 'taskloop', 'simd'])
         if end_loop:
             errors.report("For loops do not require an end clause. This clause is ignored",
@@ -1863,34 +1896,34 @@ class SemanticParser(BasicParser):
             expr.clear_user_nodes()
             return expr
 
-    def _visit_Literal(self, expr, **settings):
+    def _visit_Literal(self, expr):
         expr.clear_user_nodes()
         return expr
-    def _visit_PythonComplex(self, expr, **settings):
+    def _visit_PythonComplex(self, expr):
         expr.clear_user_nodes()
         return expr
-    def _visit_Pass(self, expr, **settings):
+    def _visit_Pass(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_Variable(self, expr, **settings):
+    def _visit_Variable(self, expr):
         name = self.scope.get_python_name(expr.name)
         return self.get_variable(name)
 
-    def _visit_str(self, expr, **settings):
+    def _visit_str(self, expr):
         return repr(expr)
 
-    def _visit_Slice(self, expr, **settings):
+    def _visit_Slice(self, expr):
         start = self._visit(expr.start) if expr.start is not None else None
         stop = self._visit(expr.stop) if expr.stop is not None else None
         step = self._visit(expr.step) if expr.step is not None else None
 
         return Slice(start, stop, step)
 
-    def _visit_IndexedElement(self, expr, **settings):
+    def _visit_IndexedElement(self, expr):
         var = self._visit(expr.base)
         # TODO check consistency of indices with shape/rank
-        args = [self._visit(idx, **settings) for idx in expr.indices]
+        args = [self._visit(idx) for idx in expr.indices]
 
         if (len(args) == 1 and isinstance(args[0], (TupleVariable, PythonTuple))):
             args = args[0]
@@ -1909,7 +1942,7 @@ class SemanticParser(BasicParser):
 
         return self._extract_indexed_from_var(var, args, expr)
 
-    def _visit_PyccelSymbol(self, expr, **settings):
+    def _visit_PyccelSymbol(self, expr):
         name = expr
 
         var = self.check_for_variable(name)
@@ -1933,7 +1966,7 @@ class SemanticParser(BasicParser):
         return var
 
 
-    def _visit_DottedName(self, expr, **settings):
+    def _visit_DottedName(self, expr):
 
         var = self.check_for_variable(_get_name(expr))
         if var:
@@ -1979,12 +2012,12 @@ class SemanticParser(BasicParser):
 
                 if isinstance(rhs, FunctionCall):
                     # If object is a function
-                    args  = self._handle_function_args(rhs.args, **settings)
+                    args  = self._handle_function_args(rhs.args)
                     func  = first[rhs_name]
                     if new_name != rhs_name:
                         if hasattr(func, 'clone') and not isinstance(func, PyccelFunctionDef):
                             func  = func.clone(new_name)
-                    return self._handle_function(expr, func, args, **settings)
+                    return self._handle_function(expr, func, args)
                 elif isinstance(rhs, Constant):
                     var = first[rhs_name]
                     if new_name != rhs_name:
@@ -2026,11 +2059,11 @@ class SemanticParser(BasicParser):
                 master = macro.master
                 args = rhs.args
                 args = [lhs] + list(args)
-                args = [self._visit(i, **settings) for i in args]
+                args = [self._visit(i) for i in args]
                 args = macro.apply(args)
                 return FunctionCall(master, args, self._current_function)
 
-            args = [self._visit(arg, **settings) for arg in
+            args = [self._visit(arg) for arg in
                     rhs.args]
             for i in methods:
                 if str(i.name) == rhs_name:
@@ -2055,7 +2088,7 @@ class SemanticParser(BasicParser):
             # standard class attribute
             if rhs in attr_name:
                 self._current_class = cls_base
-                second = self._visit(rhs, **settings)
+                second = self._visit(rhs)
                 self._current_class = None
                 return second.clone(second.name, new_class = DottedVariable, lhs = visited_lhs)
 
@@ -2089,12 +2122,12 @@ class SemanticParser(BasicParser):
             bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
             severity='fatal')
 
-    def _visit_PyccelOperator(self, expr, **settings):
-        args     = [self._visit(a, **settings) for a in expr.args]
+    def _visit_PyccelOperator(self, expr):
+        args     = [self._visit(a) for a in expr.args]
         return self._create_PyccelOperator(expr, args)
 
-    def _visit_PyccelAdd(self, expr, **settings):
-        args = [self._visit(a, **settings) for a in expr.args]
+    def _visit_PyccelAdd(self, expr):
+        args = [self._visit(a) for a in expr.args]
         if isinstance(args[0], (TupleVariable, PythonTuple, Concatenate, Duplicate)):
             is_homogeneous = all((isinstance(a, (TupleVariable, PythonTuple)) and a.is_homogeneous) \
                                 or isinstance(a, (Concatenate, Duplicate)) for a in args)
@@ -2121,8 +2154,8 @@ class SemanticParser(BasicParser):
             expr_new = self._create_PyccelOperator(expr, args)
         return expr_new
 
-    def _visit_PyccelMul(self, expr, **settings):
-        args = [self._visit(a, **settings) for a in expr.args]
+    def _visit_PyccelMul(self, expr):
+        args = [self._visit(a) for a in expr.args]
         if isinstance(args[0], (TupleVariable, PythonTuple, PythonList)):
             expr_new = self._create_Duplicate(args[0], args[1])
         elif isinstance(args[1], (TupleVariable, PythonTuple, PythonList)):
@@ -2131,8 +2164,8 @@ class SemanticParser(BasicParser):
             expr_new = self._create_PyccelOperator(expr, args)
         return expr_new
 
-    def _visit_PyccelPow(self, expr, **settings):
-        base, exponent = [self._visit(a, **settings) for a in expr.args]
+    def _visit_PyccelPow(self, expr):
+        base, exponent = [self._visit(a) for a in expr.args]
 
         exp_val = exponent
         if isinstance(exponent, LiteralInteger):
@@ -2163,9 +2196,9 @@ class SemanticParser(BasicParser):
         else:
             return PyccelPow(base, exponent)
 
-    def _visit_MathSqrt(self, expr, **settings):
+    def _visit_MathSqrt(self, expr):
         func = self.scope.find(expr.funcdef, 'functions')
-        arg, = self._handle_function_args(expr.args, **settings) #pylint: disable=unbalanced-tuple-unpacking
+        arg, = self._handle_function_args(expr.args) #pylint: disable=unbalanced-tuple-unpacking
         if isinstance(arg.value, PyccelMul):
             mul1, mul2 = arg.value.args
             mul1_syn, mul2_syn = expr.args[0].value.args
@@ -2218,9 +2251,9 @@ class SemanticParser(BasicParser):
 
                 return self._visit(new_call)
 
-        return self._handle_function(expr, func, (arg,), **settings)
+        return self._handle_function(expr, func, (arg,))
 
-    def _visit_Lambda(self, expr, **settings):
+    def _visit_Lambda(self, expr):
         expr_names = set(str(a) for a in expr.expr.get_attribute_nodes(PyccelSymbol))
         var_names = map(str, expr.variables)
         missing_vars = expr_names.difference(var_names)
@@ -2243,7 +2276,7 @@ class SemanticParser(BasicParser):
                 expr = Lambda(tuple(expr.variables), expr_new)
         return expr
 
-    def _visit_FunctionCall(self, expr, **settings):
+    def _visit_FunctionCall(self, expr):
         name     = expr.funcdef
         try:
             name = self.scope.get_expected_name(name)
@@ -2256,9 +2289,9 @@ class SemanticParser(BasicParser):
         if isinstance(func, PyccelFunctionDef):
             annotation_method = '_visit_' + func.cls_name.__name__
             if hasattr(self, annotation_method):
-                return getattr(self, annotation_method)(expr, **settings)
+                return getattr(self, annotation_method)(expr)
 
-        args = self._handle_function_args(expr.args, **settings)
+        args = self._handle_function_args(expr.args)
         # Correct keyword names if scope is available
         # The scope is only available if the function body has been parsed
         # (i.e. not for headers or builtin functions)
@@ -2318,15 +2351,15 @@ class SemanticParser(BasicParser):
                         bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
                         severity='fatal')
             else:
-                return self._handle_function(expr, func, args, **settings)
+                return self._handle_function(expr, func, args)
 
-    def _visit_Expr(self, expr, **settings):
+    def _visit_Expr(self, expr):
         errors.report(PYCCEL_RESTRICTION_TODO, symbol=expr,
             bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
             severity='fatal')
 
 
-    def _visit_Assign(self, expr, **settings):
+    def _visit_Assign(self, expr):
         # TODO unset position at the end of this part
         new_expressions = []
         fst = expr.fst
@@ -2338,7 +2371,7 @@ class SemanticParser(BasicParser):
         # Steps before visiting
         if isinstance(rhs, GeneratorComprehension):
             rhs.substitute(rhs.lhs, lhs)
-            genexp = self._assign_GeneratorComprehension(_get_name(lhs), rhs, **settings)
+            genexp = self._assign_GeneratorComprehension(_get_name(lhs), rhs)
             if isinstance(expr, AugAssign):
                 new_expressions.append(genexp)
                 rhs = genexp.lhs
@@ -2348,7 +2381,7 @@ class SemanticParser(BasicParser):
                 new_expressions.append(genexp)
                 rhs = genexp.lhs
         elif isinstance(rhs, IfTernaryOperator):
-            value_true  = self._visit(rhs.value_true, **settings)
+            value_true  = self._visit(rhs.value_true)
             if value_true.rank > 0 or value_true.dtype is NativeString():
                 # Temporarily deactivate type checks to construct syntactic assigns
                 pyccel_stage.set_stage('syntactic')
@@ -2356,7 +2389,7 @@ class SemanticParser(BasicParser):
                 assign_false = Assign(lhs, rhs.value_false, fst = fst)
                 pyccel_stage.set_stage('semantic')
 
-                cond  = self._visit(rhs.cond, **settings)
+                cond  = self._visit(rhs.cond)
                 true_section  = IfSection(cond, [self._visit(assign_true)])
                 false_section = IfSection(LiteralTrue(), [self._visit(assign_false)])
                 return If(true_section, false_section)
@@ -2366,13 +2399,13 @@ class SemanticParser(BasicParser):
             name = rhs.funcdef
             macro = self.scope.find(name, 'macros')
             if macro is None:
-                rhs = self._visit(rhs, **settings)
+                rhs = self._visit(rhs)
             else:
 
                 # TODO check types from FunctionDef
                 master = macro.master
                 results = []
-                args = [self._visit(i, **settings) for i in rhs.args]
+                args = [self._visit(i) for i in rhs.args]
                 args_names = [arg.value.name for arg in args if isinstance(arg.value, Variable)]
                 d_m_args = {arg.value.name:arg.value for arg in macro.master_arguments
                                   if isinstance(arg.value, Variable)}
@@ -2384,7 +2417,7 @@ class SemanticParser(BasicParser):
                     if m_result in d_m_args and not result in args_names:
                         d_result = self._infer_type(d_m_args[m_result])
                         d_result['shape'] = shape
-                        tmp = self._assign_lhs_variable(result, d_result, None, new_expressions, False, **settings)
+                        tmp = self._assign_lhs_variable(result, d_result, None, new_expressions, False)
                         results.append(tmp)
                     elif result in args_names:
                         _name = _get_name(result)
@@ -2416,7 +2449,7 @@ class SemanticParser(BasicParser):
             name = _get_name(var)
             macro = self.scope.find(name, 'macros')
             if macro is None:
-                rhs = self._visit(rhs, **settings)
+                rhs = self._visit(rhs)
             else:
                 master = macro.master
                 if isinstance(macro, MacroVariable):
@@ -2424,7 +2457,7 @@ class SemanticParser(BasicParser):
                 else:
                     # If macro is function, create left-hand side variable
                     if isinstance(master, FunctionDef) and master.results:
-                        d_var = self._infer_type(master.results[0].var, **settings)
+                        d_var = self._infer_type(master.results[0].var)
                         dtype = d_var.pop('datatype')
                         lhs = Variable(dtype, lhs.name, **d_var, is_temp=lhs.is_temp)
                         var = self.check_for_variable(lhs.name)
@@ -2442,7 +2475,7 @@ class SemanticParser(BasicParser):
 
                     args = rhs.rhs.args
                     args = [rhs.lhs] + list(args)
-                    args = [self._visit(i, **settings) for i in args]
+                    args = [self._visit(i) for i in args]
 
                     args = macro.apply(args, results=results)
 
@@ -2453,7 +2486,7 @@ class SemanticParser(BasicParser):
                         return FunctionCall(master, args, self._current_function)
 
         else:
-            rhs = self._visit(rhs, **settings)
+            rhs = self._visit(rhs)
 
         if isinstance(rhs, FunctionDef):
 
@@ -2462,7 +2495,6 @@ class SemanticParser(BasicParser):
             rhs = rhs.rename(expr.lhs.name)
             for i in rhs.body:
                 i.set_fst(fst)
-            rhs = self._visit_FunctionDef(rhs, **settings)
             return rhs
 
         elif isinstance(rhs, CodeBlock):
@@ -2478,7 +2510,7 @@ class SemanticParser(BasicParser):
             if isinstance(lhs, PyccelSymbol):
                 name = lhs
                 if self.check_for_variable(name) is None:
-                    d_var = self._infer_type(stmt, **settings)
+                    d_var = self._infer_type(stmt)
                     dtype = d_var.pop('datatype')
                     lhs = Variable(dtype, name , **d_var, is_temp = lhs.is_temp)
                     self.scope.insert_variable(lhs)
@@ -2497,9 +2529,9 @@ class SemanticParser(BasicParser):
                 results = func.results
                 if results:
                     if len(results)==1:
-                        d_var = self._infer_type(results[0].var, **settings)
+                        d_var = self._infer_type(results[0].var)
                     else:
-                        d_var = self._infer_type(PythonTuple(*[r.var for r in results]), **settings)
+                        d_var = self._infer_type(PythonTuple(*[r.var for r in results]))
                 elif expr.lhs.is_temp:
                     return rhs
                 else:
@@ -2526,7 +2558,7 @@ class SemanticParser(BasicParser):
                         d_var['order'          ] = arg.order
 
             elif isinstance(func, Interface):
-                d_var = [self._infer_type(i.var, **settings) for i in
+                d_var = [self._infer_type(i.var) for i in
                          func.functions[0].results]
 
                 # TODO imporve this will not work for
@@ -2534,10 +2566,10 @@ class SemanticParser(BasicParser):
                 d_var[0]['datatype'] = rhs.dtype
 
             else:
-                d_var = self._infer_type(rhs, **settings)
+                d_var = self._infer_type(rhs)
 
         elif isinstance(rhs, NumpyTranspose):
-            d_var  = self._infer_type(rhs, **settings)
+            d_var  = self._infer_type(rhs)
             if d_var['memory_handling'] == 'alias' and not isinstance(lhs, IndexedElement):
                 rhs = rhs.internal_var
         elif isinstance(rhs, PyccelInternalFunction) and isinstance(rhs.dtype, NativeVoid):
@@ -2547,7 +2579,7 @@ class SemanticParser(BasicParser):
                 raise NotImplementedError("Cannot assign result of a function without a return")
 
         else:
-            d_var  = self._infer_type(rhs, **settings)
+            d_var  = self._infer_type(rhs)
             d_list = d_var if isinstance(d_var, list) else [d_var]
 
             for d in d_list:
@@ -2580,7 +2612,7 @@ class SemanticParser(BasicParser):
                         bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
                         severity='error')
                     return None
-            lhs = self._assign_lhs_variable(lhs, d_var, rhs, new_expressions, isinstance(expr, AugAssign), **settings)
+            lhs = self._assign_lhs_variable(lhs, d_var, rhs, new_expressions, isinstance(expr, AugAssign))
         elif isinstance(lhs, PythonTuple):
             n = len(lhs)
             if isinstance(rhs, (PythonTuple, InhomogeneousTupleVariable, FunctionCall)):
@@ -2590,8 +2622,8 @@ class SemanticParser(BasicParser):
                     r_iter = rhs
                 new_lhs = []
                 for i,(l,r) in enumerate(zip(lhs,r_iter)):
-                    d = self._infer_type(r, **settings)
-                    new_lhs.append( self._assign_lhs_variable(l, d, r, new_expressions, isinstance(expr, AugAssign),arr_in_multirets=r.rank>0 ,**settings) )
+                    d = self._infer_type(r)
+                    new_lhs.append( self._assign_lhs_variable(l, d, r, new_expressions, isinstance(expr, AugAssign),arr_in_multirets=r.rank>0 ) )
                 lhs = PythonTuple(*new_lhs)
 
             elif isinstance(rhs, HomogeneousTupleVariable):
@@ -2600,7 +2632,7 @@ class SemanticParser(BasicParser):
                 new_rhs = []
                 for i,l in enumerate(lhs):
                     new_lhs.append( self._assign_lhs_variable(l, d_var.copy(),
-                        rhs[i], new_expressions, isinstance(expr, AugAssign), **settings) )
+                        rhs[i], new_expressions, isinstance(expr, AugAssign)) )
                     new_rhs.append(rhs[i])
                 rhs = PythonTuple(*new_rhs)
                 d_var = [d_var]
@@ -2610,10 +2642,10 @@ class SemanticParser(BasicParser):
                 new_lhs = []
                 if hasattr(rhs,'__getitem__'):
                     for i,l in enumerate(lhs):
-                        new_lhs.append( self._assign_lhs_variable(l, d_var[i].copy(), rhs[i], new_expressions, isinstance(expr, AugAssign), **settings) )
+                        new_lhs.append( self._assign_lhs_variable(l, d_var[i].copy(), rhs[i], new_expressions, isinstance(expr, AugAssign)) )
                 else:
                     for i,l in enumerate(lhs):
-                        new_lhs.append( self._assign_lhs_variable(l, d_var[i].copy(), rhs, new_expressions, isinstance(expr, AugAssign), **settings) )
+                        new_lhs.append( self._assign_lhs_variable(l, d_var[i].copy(), rhs, new_expressions, isinstance(expr, AugAssign)) )
                 lhs = PythonTuple(*new_lhs)
 
             elif d_var['shape'][0]==n:
@@ -2621,7 +2653,7 @@ class SemanticParser(BasicParser):
                 new_rhs = []
 
                 for l, r in zip(lhs, rhs):
-                    new_lhs.append( self._assign_lhs_variable(l, self._infer_type(r), r, new_expressions, isinstance(expr, AugAssign), **settings) )
+                    new_lhs.append( self._assign_lhs_variable(l, self._infer_type(r), r, new_expressions, isinstance(expr, AugAssign)) )
                     new_rhs.append(r)
 
                 lhs = PythonTuple(*new_lhs)
@@ -2632,7 +2664,7 @@ class SemanticParser(BasicParser):
                     severity='error')
                 return None
         else:
-            lhs = self._visit(lhs, **settings)
+            lhs = self._visit(lhs)
 
         if not isinstance(lhs, (list, tuple)):
             lhs = [lhs]
@@ -2709,12 +2741,12 @@ class SemanticParser(BasicParser):
             result = CodeBlock(new_expressions)
             return result
 
-    def _visit_For(self, expr, **settings):
+    def _visit_For(self, expr):
 
         scope = self.create_new_loop_scope()
 
         # treatment of the index/indices
-        iterable = Iterable(self._visit(expr.iterable, **settings))
+        iterable = Iterable(self._visit(expr.iterable))
 
         new_expr = []
 
@@ -2737,7 +2769,7 @@ class SemanticParser(BasicParser):
             if index is None:
                 index = self._assign_lhs_variable(syntactic_index, iterator_d_var,
                                 rhs=start, new_expressions=new_expr,
-                                is_augassign=False, **settings)
+                                is_augassign=False)
             iterable.set_loop_counter(index)
 
         if isinstance(iterator, PyccelSymbol):
@@ -2746,13 +2778,13 @@ class SemanticParser(BasicParser):
 
             target = self._assign_lhs_variable(iterator, iterator_d_var,
                             rhs=iterator_rhs, new_expressions=new_expr,
-                            is_augassign=False, **settings)
+                            is_augassign=False)
 
         elif isinstance(iterator, PythonTuple):
             iterator_rhs = iterable.get_target_from_range()
             target = [self._assign_lhs_variable(it, self._infer_type(rhs),
                                 rhs=rhs, new_expressions=new_expr,
-                                is_augassign=False, **settings)
+                                is_augassign=False)
                         for it, rhs in zip(iterator, iterator_rhs)]
         else:
 
@@ -2761,7 +2793,7 @@ class SemanticParser(BasicParser):
                    severity='error')
 
 
-        body = self._visit(expr.body, **settings)
+        body = self._visit(expr.body)
 
         self.exit_loop_scope()
 
@@ -2780,7 +2812,7 @@ class SemanticParser(BasicParser):
         return for_expr
 
 
-    def _visit_FunctionalFor(self, expr, **settings):
+    def _visit_FunctionalFor(self, expr):
         old_index   = expr.index
         new_index   = self.scope.get_new_name()
         expr.substitute(old_index, new_index)
@@ -2804,15 +2836,15 @@ class SemanticParser(BasicParser):
             step  = LiteralInteger(1)
             var   = indices[i]
             i += 1
-            a     = self._visit(body.iterable, **settings)
+            a     = self._visit(body.iterable)
             if isinstance(a, PythonRange):
                 var   = self._create_variable(var, 'int', start, {})
-                dvar  = self._infer_type(var, **settings)
+                dvar  = self._infer_type(var)
                 stop  = a.stop
                 start = a.start
                 step  = a.step
             elif isinstance(a, (PythonZip, PythonEnumerate)):
-                dvar  = self._infer_type(a.element, **settings)
+                dvar  = self._infer_type(a.element)
                 dtype = dvar.pop('datatype')
                 if dvar['rank'] > 0:
                     dvar['rank' ] -= 1
@@ -2822,7 +2854,7 @@ class SemanticParser(BasicParser):
                 var  = Variable(dtype, var, **dvar)
                 stop = a.element.shape[0]
             elif isinstance(a, Variable):
-                dvar  = self._infer_type(a, **settings)
+                dvar  = self._infer_type(a)
                 dtype = dvar.pop('datatype')
                 if dvar['rank'] == 1:
                     dvar['rank']  = 0
@@ -2841,7 +2873,7 @@ class SemanticParser(BasicParser):
                               severity='fatal')
             existing_var = self.scope.find(var.name, 'variables')
             if existing_var:
-                if self._infer_type(existing_var, **settings) != dvar:
+                if self._infer_type(existing_var) != dvar:
                     errors.report(f"Variable {var} already exists with different type",
                             symbol = expr, severity='error')
             else:
@@ -2925,8 +2957,8 @@ class SemanticParser(BasicParser):
 
         # we annotate the target to infere the type of the list created
 
-        target = self._visit(target, **settings)
-        d_var = self._infer_type(target, **settings)
+        target = self._visit(target)
+        d_var = self._infer_type(target)
 
         dtype = d_var['datatype']
 
@@ -2951,7 +2983,7 @@ class SemanticParser(BasicParser):
         # ...
         lhs_symbol = expr.lhs.base
         ne = []
-        lhs = self._assign_lhs_variable(lhs_symbol, d_var, rhs=expr, new_expressions=ne, is_augassign=False, **settings)
+        lhs = self._assign_lhs_variable(lhs_symbol, d_var, rhs=expr, new_expressions=ne, is_augassign=False)
         lhs_alloc = ne[0]
 
         if isinstance(target, PythonTuple) and not target.is_homogeneous:
@@ -2961,8 +2993,8 @@ class SemanticParser(BasicParser):
 
         target.invalidate_node()
 
-        loops = [self._visit(i, **settings) for i in expr.loops]
-        index = self._visit(index, **settings)
+        loops = [self._visit(i) for i in expr.loops]
+        index = self._visit(index)
 
         l = loops[-1]
         for idx in indices:
@@ -2977,7 +3009,7 @@ class SemanticParser(BasicParser):
 
         return CodeBlock([lhs_alloc, FunctionalFor(loops, lhs=lhs, indices=indices, index=index)])
 
-    def _visit_GeneratorComprehension(self, expr, **settings):
+    def _visit_GeneratorComprehension(self, expr):
         lhs = self.check_for_variable(expr.lhs)
         if lhs is None:
             if expr.lhs.is_temp:
@@ -2991,16 +3023,16 @@ class SemanticParser(BasicParser):
         else:
             return lhs
 
-    def _visit_While(self, expr, **settings):
+    def _visit_While(self, expr):
 
         scope = self.create_new_loop_scope()
-        test = self._visit(expr.test, **settings)
-        body = self._visit(expr.body, **settings)
+        test = self._visit(expr.test)
+        body = self._visit(expr.body)
         self.exit_loop_scope()
 
         return While(test, body, scope=scope)
 
-    def _visit_IfSection(self, expr, **settings):
+    def _visit_IfSection(self, expr):
         condition = expr.condition
 
         name_symbol = PyccelSymbol('__name__')
@@ -3023,8 +3055,8 @@ class SemanticParser(BasicParser):
 
         return IfSection(cond, body)
 
-    def _visit_If(self, expr, **settings):
-        args = [self._visit(i, **settings) for i in expr.blocks]
+    def _visit_If(self, expr):
+        args = [self._visit(i) for i in expr.blocks]
 
         conds = [b.condition for b in args]
         if any(isinstance(c, InProgram) for c in conds):
@@ -3051,8 +3083,8 @@ class SemanticParser(BasicParser):
 
         return If(*args)
 
-    def _visit_IfTernaryOperator(self, expr, **settings):
-        value_true  = self._visit(expr.value_true, **settings)
+    def _visit_IfTernaryOperator(self, expr):
+        value_true  = self._visit(expr.value_true)
         if value_true.rank > 0 or value_true.dtype is NativeString():
             lhs = PyccelSymbol(self.scope.get_new_name(), is_temp=True)
             # Temporarily deactivate type checks to construct syntactic assigns
@@ -3061,18 +3093,18 @@ class SemanticParser(BasicParser):
             assign_false = Assign(lhs, expr.value_false, fst = expr.fst)
             pyccel_stage.set_stage('semantic')
 
-            cond  = self._visit(expr.cond, **settings)
+            cond  = self._visit(expr.cond)
             true_section  = IfSection(cond, [self._visit(assign_true)])
             false_section = IfSection(LiteralTrue(), [self._visit(assign_false)])
             self._additional_exprs[-1].append(If(true_section, false_section))
 
             return self._visit(lhs)
         else:
-            cond        = self._visit(expr.cond, **settings)
-            value_false = self._visit(expr.value_false, **settings)
+            cond        = self._visit(expr.cond)
+            value_false = self._visit(expr.value_false)
             return IfTernaryOperator(cond, value_true, value_false)
 
-    def _visit_VariableHeader(self, expr, **settings):
+    def _visit_VariableHeader(self, expr):
 
         # TODO improve
         #      move it to the ast like create_definition for FunctionHeader?
@@ -3086,24 +3118,24 @@ class SemanticParser(BasicParser):
         self.scope.insert_variable(var)
         return expr
 
-    def _visit_FunctionHeader(self, expr, **settings):
+    def _visit_FunctionHeader(self, expr):
         # TODO should we return it and keep it in the AST?
         expr.clear_user_nodes()
         self.scope.insert_header(expr)
         return expr
 
-    def _visit_Template(self, expr, **settings):
+    def _visit_Template(self, expr):
         expr.clear_user_nodes()
         self.scope.insert_template(expr)
         return expr
 
-    def _visit_ClassHeader(self, expr, **settings):
+    def _visit_ClassHeader(self, expr):
         # TODO should we return it and keep it in the AST?
         expr.clear_user_nodes()
         self.scope.insert_header(expr)
         return expr
 
-    def _visit_Return(self, expr, **settings):
+    def _visit_Return(self, expr):
 
         results     = expr.expr
         f_name      = self._current_function
@@ -3118,7 +3150,7 @@ class SemanticParser(BasicParser):
                 a = self._visit(Assign(v, r, fst=expr.fst))
                 assigns.append(a)
 
-        results = [self._visit(i.var, **settings) for i in return_objs]
+        results = [self._visit(i.var) for i in return_objs]
 
         # add the Deallocate node before the Return node and eliminating the Deallocate nodes
         # the arrays that will be returned.
@@ -3129,7 +3161,7 @@ class SemanticParser(BasicParser):
             expr  = Return(results)
         return expr
 
-    def _visit_FunctionDef(self, expr, **settings):
+    def _visit_FunctionDef(self, expr):
         name            = self.scope.get_expected_name(expr.name)
         cls_name        = expr.cls_name
         decorators      = expr.decorators
@@ -3261,7 +3293,7 @@ class SemanticParser(BasicParser):
                         a_new = FunctionAddress(self.scope.get_expected_name(a.name),
                                         ahv.arguments, ahv.results, [], **d_var)
                     else:
-                        d_var = self._infer_type(ahv, **settings)
+                        d_var = self._infer_type(ahv)
                         d_var['shape'] = ahv.alloc_shape
                         d_var['is_argument'] = True
                         d_var['is_const'] = ahv.is_const
@@ -3302,7 +3334,7 @@ class SemanticParser(BasicParser):
 
                 for a, ah in zip(results, header_results):
                     av = a.var
-                    d_var = self._infer_type(ah.var, **settings)
+                    d_var = self._infer_type(ah.var)
                     dtype = d_var.pop('datatype')
                     a_new = Variable(dtype, self.scope.get_expected_name(av),
                             **d_var, is_temp = av.is_temp)
@@ -3450,7 +3482,7 @@ class SemanticParser(BasicParser):
             self.insert_function(funcs)
 #        TODO move this to codegen
 #        if vec_func:
-#           self._visit_FunctionDef(vec_func, **settings)
+#           self._visit_FunctionDef(vec_func)
 #           vec_func = self.scope.functions.pop(vec_name)
 #           if isinstance(funcs, Interface):
 #               funcs = list(funcs.funcs)+[vec_func]
@@ -3504,7 +3536,7 @@ class SemanticParser(BasicParser):
         else:
             return PythonPrint(args)
 
-    def _visit_ClassDef(self, expr, **settings):
+    def _visit_ClassDef(self, expr):
 
         # TODO - improve the use and def of interfaces
         #      - wouldn't be better if it is done inside ClassDef?
@@ -3526,7 +3558,7 @@ class SemanticParser(BasicParser):
             m_name = method.name.replace("'", '')
 
             if m_name == '__init__':
-                self._visit_FunctionDef(method, **settings)
+                self._visit_FunctionDef(method)
                 methods.pop(i)
                 const = self.scope.functions.pop(m_name)
                 break
@@ -3540,7 +3572,7 @@ class SemanticParser(BasicParser):
 
         ms = []
         for i in methods:
-            self._visit_FunctionDef(i, **settings)
+            self._visit_FunctionDef(i)
             m_name = i.name.replace("'", '')
             m = self.scope.functions.pop(m_name)
             ms.append(m)
@@ -3568,12 +3600,12 @@ class SemanticParser(BasicParser):
 
         return EmptyNode()
 
-    def _visit_Del(self, expr, **settings):
+    def _visit_Del(self, expr):
 
-        ls = [self._visit(i, **settings) for i in expr.variables]
+        ls = [self._visit(i) for i in expr.variables]
         return Del(ls)
 
-    def _visit_PyccelIs(self, expr, **settings):
+    def _visit_PyccelIs(self, expr):
         # Handles PyccelIs and PyccelIsNot
         IsClass = type(expr)
 
@@ -3625,7 +3657,7 @@ class SemanticParser(BasicParser):
             severity='error')
         return IsClass(var1, var2)
 
-    def _visit_Import(self, expr, **settings):
+    def _visit_Import(self, expr):
 
         # TODO - must have a dict where to store things that have been
         #        imported
@@ -3771,24 +3803,24 @@ class SemanticParser(BasicParser):
 
 
 
-    def _visit_With(self, expr, **settings):
+    def _visit_With(self, expr):
         scope = self.create_new_loop_scope()
 
-        domaine = self._visit(expr.test, **settings)
+        domaine = self._visit(expr.test)
         parent  = domaine.cls_base
         if not parent.is_with_construct:
             errors.report(UNDEFINED_WITH_ACCESS,
                    bounding_box=(self._current_fst_node.lineno, self._current_fst_node.col_offset),
                    severity='fatal')
 
-        body = self._visit(expr.body, **settings)
+        body = self._visit(expr.body)
 
         self.exit_loop_scope()
         return With(domaine, body, scope).block
 
 
 
-    def _visit_MacroFunction(self, expr, **settings):
+    def _visit_MacroFunction(self, expr):
         # we change here the master name to its FunctionDef
 
         f_name = expr.master
@@ -3826,11 +3858,11 @@ class SemanticParser(BasicParser):
 
         return macro
 
-    def _visit_MacroShape(self, expr, **settings):
+    def _visit_MacroShape(self, expr):
         expr.clear_user_nodes()
         return expr
 
-    def _visit_MacroVariable(self, expr, **settings):
+    def _visit_MacroVariable(self, expr):
 
         master = expr.master
         if isinstance(master, DottedName):
@@ -3849,13 +3881,13 @@ class SemanticParser(BasicParser):
         self.scope.insert_macro(expr)
         return expr
 
-    def _visit_StarredArguments(self, expr, **settings):
+    def _visit_StarredArguments(self, expr):
         var = self._visit(expr.args_var)
         assert(var.rank==1)
         size = var.shape[0]
         return StarredArguments([var[i] for i in range(size)])
 
-    def _visit_NumpyMatmul(self, expr, **settings):
+    def _visit_NumpyMatmul(self, expr):
         if isinstance(expr, FunctionCall):
             a = self._visit(expr.args[0].value)
             b = self._visit(expr.args[1].value)
@@ -3865,12 +3897,12 @@ class SemanticParser(BasicParser):
             b = self._visit(expr.b)
         return NumpyMatmul(a, b)
 
-    def _visit_Assert(self, expr, **settings):
-        test = self._visit(expr.test, **settings)
+    def _visit_Assert(self, expr):
+        test = self._visit(expr.test)
         return Assert(test)
 
-    def _visit_NumpyWhere(self, func_call, **settings):
-        func_call_args = self._handle_function_args(func_call.args, **settings)
+    def _visit_NumpyWhere(self, func_call):
+        func_call_args = self._handle_function_args(func_call.args)
         # expr is a FunctionCall
         args = [a.value for a in func_call_args if not a.has_keyword]
         kwargs = {a.keyword: a.value for a in func_call.args if a.has_keyword}
@@ -3879,8 +3911,8 @@ class SemanticParser(BasicParser):
             return self._visit_NumpyNonZero(func_call)
         return NumpyWhere(*args, **kwargs)
 
-    def _visit_NumpyNonZero(self, func_call, **settings):
-        func_call_args = self._handle_function_args(func_call.args, **settings)
+    def _visit_NumpyNonZero(self, func_call):
+        func_call_args = self._handle_function_args(func_call.args)
         # expr is a FunctionCall
         arg = func_call_args[0].value
         if not isinstance(arg, Variable):
@@ -3890,6 +3922,6 @@ class SemanticParser(BasicParser):
             arg = self._visit(new_symbol)
         return NumpyWhere(arg)
 
-    def _visit_FunctionDefResult(self, expr, **settings):
+    def _visit_FunctionDefResult(self, expr):
         var = self._visit(expr.var)
         return FunctionDefResult(var, annotation = expr.annotation)