+++ CHANGELOG.md
@@ -91,5 +91,6 @@ All notable changes to this project will be documented in this file.
 -   #1245 : Reduce false negative test results by using a tolerance to compare floats.
 -   #1272 : Remove use of deprecated NumPy syntax in tests.
 -   #1253 : Provide minimum requirements.
+-   \[INTERNALS\]  #1385 : Remove unused settings keyword arguments from `_visit` function.
 
 ### Deprecated